Move navbar components into a folder
