+++ components/PageTemplate.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import { useRouter } from 'next/router';
 import { FunctionComponent, ReactNode } from 'react';
 import Footer from './Footer';
-import Navbar from './Navbar';
+import Navbar from './Navbar/Navbar';
 
 type Props = {
   children?: ReactNode;
